backend/requirements.txt
fastapi
uvicorn[standard]
gunicorn
databases[postgresql]
asyncpg
psycopg2-binary
python-multipart
pydantic
boto3
aiosqlite
python-jose[cryptography]
bcrypt
httpx
